[Bronze II] Title: 피보나치 수, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/2747. 피보나치 수/피보나치 수.cc
@@ -1,6 +1,6 @@
 #include <iostream>
 using namespace std;
-long dp[45];
+int dp[45];
 
 int main() {
 	ios::sync_with_stdio(false);
